refactor: KDA-42 result extension 이동
