+++ domain/src/main/java/com/upf464/koonsdiary/domain/usecase/user/SignInWithUsernameUseCase.kt
@@ -1,7 +1,7 @@
 package com.upf464.koonsdiary.domain.usecase.user
 
+import com.upf464.koonsdiary.common.extension.flatMap
 import com.upf464.koonsdiary.domain.common.HashGenerator
-import com.upf464.koonsdiary.domain.common.flatMap
 import com.upf464.koonsdiary.domain.repository.MessageRepository
 import com.upf464.koonsdiary.domain.repository.UserRepository
 import com.upf464.koonsdiary.domain.request.user.SignInWithUsernameRequest
