+++ domain/src/main/java/com/upf464/koonsdiary/domain/usecase/user/SignUpWithKakaoUseCase.kt
@@ -1,7 +1,7 @@
 package com.upf464.koonsdiary.domain.usecase.user
 
-import com.upf464.koonsdiary.domain.common.flatMap
-import com.upf464.koonsdiary.domain.common.handleWith
+import com.upf464.koonsdiary.common.extension.flatMap
+import com.upf464.koonsdiary.common.extension.handleWith
 import com.upf464.koonsdiary.domain.error.SignInError
 import com.upf464.koonsdiary.domain.model.User
 import com.upf464.koonsdiary.domain.repository.MessageRepository
