+++ domain/src/main/java/com/upf464/koonsdiary/domain/usecase/user/SignUpWithUsernameUseCase.kt
@@ -1,8 +1,8 @@
 package com.upf464.koonsdiary.domain.usecase.user
 
+import com.upf464.koonsdiary.common.extension.flatMap
 import com.upf464.koonsdiary.domain.common.HashGenerator
 import com.upf464.koonsdiary.domain.common.UserValidator
-import com.upf464.koonsdiary.domain.common.flatMap
 import com.upf464.koonsdiary.domain.error.SignUpError
 import com.upf464.koonsdiary.domain.model.User
 import com.upf464.koonsdiary.domain.repository.MessageRepository
