+++ kakao/src/main/java/com/upf464/koonsdiary/kakao/service/KakaoServiceImpl.kt
@@ -6,7 +6,7 @@ import com.kakao.sdk.common.model.AuthError
 import com.kakao.sdk.common.model.ClientError
 import com.kakao.sdk.common.model.ClientErrorCause
 import com.kakao.sdk.user.UserApiClient
-import com.upf464.koonsdiary.domain.common.errorMap
+import com.upf464.koonsdiary.common.extension.errorMap
 import com.upf464.koonsdiary.domain.error.SignInError
 import com.upf464.koonsdiary.domain.service.KakaoService
 import com.upf464.koonsdiary.kakao.mapper.toDomain
